requirements.txt
Django>=2.2.13,<2.3.0
djangorestframework>=3.9.0,<3.10.0
psycopg2>=2.7.5,<2.8.0
Pillow>=7.1.0,<8.0.0
coverage>=5.3,<6.0.0

flake8>=3.8.3,<3.9.0
